requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
googlemaps==4.10.0
anthropic>=0.25.0
